+++ app/components/primer/alpha/action_list.rb
@@ -53,7 +53,7 @@ module Primer
       #
       # @param system_arguments [Hash] The arguments accepted by <%= link_to_component(Primer::Alpha::ActionList::Heading) %>.
       renders_one :heading, lambda { |**system_arguments|
-        Heading.new(list_id: @id, **system_arguments)
+        Heading.new(**system_arguments)
       }
 
       # Items.
@@ -74,7 +74,7 @@ module Primer
         set_slot(:items, { renderable: Divider, collection: true }, **system_arguments, &block)
       end
 
-      attr_reader :select_variant, :role
+      attr_reader :id, :select_variant, :role
 
       # @param id [String] HTML ID value.
       # @param role [Boolean] ARIA role describing the function of the list. listbox and menu are a common values.
@@ -82,6 +82,7 @@ module Primer
       # @param scheme [Symbol] <%= one_of(Primer::Alpha::ActionList::SCHEME_OPTIONS) %> `inset` children are offset (vertically and horizontally) from list edges. `full` (default) children are flush (vertically and horizontally) with list edges.
       # @param show_dividers [Boolean] Display a divider above each item in the list when it does not follow a header or divider.
       # @param select_variant [Symbol] How items may be selected in the list. <%= one_of(Primer::Alpha::ActionList::SELECT_VARIANT_OPTIONS) %>
+      # @param form_arguments [Hash] Allows an `ActionList` to act as a select list in multi- and single-select modes. Pass the `builder:` and `name:` options to this hash. `builder:` should be an instance of `ActionView::Helpers::FormBuilder`, which are created by the standard Rails `#form_with` and `#form_for` helpers. The `name:` option is the desired name of the field that will be included in the params sent to the server on form submission. *NOTE*: Consider using an <%= link_to_component(Primer::Alpha::ActionMenu) %> instead of using this feature directly.
       # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
       def initialize(
         id: self.class.generate_id,
@@ -90,6 +91,7 @@ module Primer
         scheme: DEFAULT_SCHEME,
         show_dividers: false,
         select_variant: DEFAULT_SELECT_VARIANT,
+        form_arguments: {},
         **system_arguments
       )
         @system_arguments = system_arguments
@@ -107,23 +109,25 @@ module Primer
           "ActionListWrap--divided" => @show_dividers
         )
 
-        @role = role || allows_selection? ? MENU_ROLE : DEFAULT_ROLE
+        @role = role || (allows_selection? ? MENU_ROLE : DEFAULT_ROLE)
         @system_arguments[:role] = @role
 
         @list_wrapper_arguments = {}
+
+        @form_builder = form_arguments[:builder]
+        @input_name = form_arguments[:name]
+
+        return unless required_form_arguments_given? && !allows_selection?
+
+        raise ArgumentError, "lists/menus that act as form inputs must also allow item selection (please pass the `select_variant:` option)"
       end
 
       # @private
       def before_render
-        aria_label = aria(:label, @system_arguments)
-        aria_labelledby = aria(:labelledby, @system_arguments)
+        return unless heading?
 
-        if heading.present?
-          @system_arguments[:"aria-labelledby"] = heading.id unless aria_labelledby
-          raise ArgumentError, "An aria-label should not be provided if a heading is present" if aria_label.present?
-        elsif aria_label.blank? && aria_labelledby.blank?
-          raise ArgumentError, "An aria-label, aria-labelledby, or heading must be provided"
-        end
+        @system_arguments[:"aria-labelledby"] = heading.title_id
+        @system_arguments[:"aria-describedby"] = heading.subtitle_id if heading.subtitle?
       end
 
       # @private
@@ -158,8 +162,24 @@ module Primer
         @system_arguments[:role] == :menu
       end
 
+      def required_form_arguments_given?
+        @form_builder && @input_name
+      end
+
+      def acts_as_form_input?
+        required_form_arguments_given? && allows_selection?
+      end
+
       # @private
       def will_add_item(_item); end
+
+      private
+
+      def with_post_list_content(&block)
+        @post_list_content_block = block
+      end
+
+      attr_reader :post_list_content_block
     end
   end
 end