+++ app/components/primer/alpha/modal_dialog.ts
@@ -12,6 +12,7 @@ const overlayStack: ModalDialogElement[] = []
 function clickHandler(event: Event) {
   const target = event.target as HTMLElement
   const button = target?.closest('button')
+
   if (!button) return
 
   // If the user is clicking a valid dialog trigger
@@ -40,6 +41,21 @@ function clickHandler(event: Event) {
     overlayStack.pop()
     topLevelDialog.close(true)
   }
+}
+
+function keydownHandler(event: Event) {
+  if (
+    !(event instanceof KeyboardEvent) ||
+    event.type !== 'keydown' ||
+    event.key !== 'Enter' ||
+    event.ctrlKey ||
+    event.altKey ||
+    event.metaKey ||
+    event.shiftKey
+  )
+    return
+
+  clickHandler(event)
 }
 
 function mousedownHandler(event: Event) {
@@ -124,6 +140,7 @@ export class ModalDialogElement extends HTMLElement {
     if (!this.hasAttribute('role')) this.setAttribute('role', 'dialog')
 
     document.addEventListener('click', clickHandler)
+    document.addEventListener('keydown', keydownHandler)
     document.addEventListener('mousedown', mousedownHandler)
 
     this.addEventListener('keydown', e => this.#keydown(e))