+++ test/lib/erblint/disallow_component_css_counter_test.rb
@@ -28,7 +28,7 @@ class DisallowComponentCssCounterTest < ErblintTestCase
       </div>
     HTML
     @linter.run(processed_source)
-    assert_includes offenses.first.message, "Primer::Alpha::Dialog or Primer::Alpha::Overlay"
+    assert_includes offenses.first.message, "Primer::Alpha::Dialog / Primer::Alpha::Overlay"
   end
 
   def test_no_warning_on_class_covered_by_others